requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
certifi==2024.12.14
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
deprecation==2.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
marshmallow==3.26.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
propcache==0.2.1
pydub==0.25.1
PyPDF2==3.0.1
python-dotenv==1.0.1
pyttsx3==2.98
simpleaudio==1.0.4
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
websockets==13.1
yarl==1.18.3
